+++ CHANGELOG.md
@@ -13,6 +13,10 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Added
+- BusinessAccountRequest resource
+- validator_link attribute to IndividualAccountRequest resource
+- complement attribute to IndividualAccountRequest address
 
 ## [0.28.0] - 2026-06-24
 ### Added